# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClinicalCompass API",
    "summary": "Authentication API with database-backed opaque tokens.",
    "description": "\n**ClinicalCompass API** \u2014 username/password authentication using opaque,\nserver-side tokens (no JWT).\n\n### Auth flow\n1. `POST /auth/register` to create a user.\n2. `POST /auth/login` with the username + password to receive an opaque\n   `access_token`.\n3. Send the token as `Authorization: Bearer <access_token>` on protected\n   endpoints.\n4. `POST /auth/logout` to revoke the token.\n\nTokens are random strings stored in the database and expire after 24 hours.\n",
    "contact": {
      "name": "ClinicalCompass",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/specialities": {
      "get": {
        "tags": [
          "specialities"
        ],
        "summary": "List specialities (paginated)",
        "description": "Returns a page of specialities. Use `limit` (1-100) and `offset` to paginate; `total` is the full count and `next`/`previous` are page URLs.",
        "operationId": "list_specialities_specialities_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search_query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive match on title or description",
              "title": "Search Query"
            },
            "description": "Case-insensitive match on title or description"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialityPage"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/specialities/count": {
      "get": {
        "tags": [
          "specialities"
        ],
        "summary": "Count of specialities",
        "description": "Returns the total number of specialities.",
        "operationId": "count_specialities_specialities_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/specialities/{speciality_id}": {
      "get": {
        "tags": [
          "specialities"
        ],
        "summary": "Get a speciality with its modules",
        "operationId": "get_speciality_specialities__speciality_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "speciality_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Speciality Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialityWithModules"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Speciality not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/specialities/{speciality_id}/modules": {
      "get": {
        "tags": [
          "specialities"
        ],
        "summary": "List modules in a speciality",
        "operationId": "list_modules_specialities__speciality_id__modules_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "speciality_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Speciality Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleOut"
                  },
                  "title": "Response List Modules Specialities  Speciality Id  Modules Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Speciality not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CountOut": {
        "properties": {
          "resource": {
            "type": "string",
            "title": "Resource",
            "examples": [
              "specialities"
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "examples": [
              42
            ]
          }
        },
        "type": "object",
        "required": [
          "resource",
          "count"
        ],
        "title": "CountOut",
        "description": "A labeled count of a resource."
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Invalid username or password"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard FastAPI error envelope."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModuleOut": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "examples": [
              "Rectal Cancer"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Staging and treatment pathways"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "speciality_id": {
            "type": "integer",
            "title": "Speciality Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "/uploads/images/abc123.png"
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "speciality_id"
        ],
        "title": "ModuleOut"
      },
...
      "SpecialityOut": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "examples": [
              "Oncology"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Cancer care pathways"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "/uploads/images/abc123.png"
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "id"
        ],
        "title": "SpecialityOut"
      },
      "SpecialityPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SpecialityOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next",
            "description": "URL of the next page, or null on the last page",
            "examples": [
              "http://localhost/specialities?limit=20&offset=20"
            ]
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous",
            "description": "URL of the previous page, or null on the first page",
            "examples": [
              null
            ]
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SpecialityPage",
        "description": "A page of specialities with next/previous page links."
      },
      "SpecialityWithModules": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "examples": [
              "Oncology"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "examples": [
              "Cancer care pathways"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "examples": [
              "/uploads/images/abc123.png"
            ]
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleOut"
            },
            "type": "array",
            "title": "Modules"
          }
        },
        "type": "object",
        "required": [
          "title",
          "id"
        ],
        "title": "SpecialityWithModules"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}